refactor(renamed checkForPorts file utilityFuncs, moved q click to utilityFuncs )
